Add explicit inferface implementation, methods from 2 interfaces
