+++ analyzers/tests/SonarAnalyzer.UnitTest/TestCases/EmptyMethod.CSharp11.cs
@@ -12,7 +12,7 @@ interface IMyInterface
 
 class MyClass : IMyInterface
 {
-    public static void StaticVirtualMethod() { } // Noncompliant
+    public static void StaticVirtualMethod() { } // Noncompliant FP
 
     public static void StaticAbstractMethod() { } // Noncompliant FP
 }
@@ -25,20 +25,4 @@ public unsafe partial class Externals
     // Code usually generated by the source generator:
     [System.Runtime.InteropServices.DllImportAttribute("ole32.dll", EntryPoint = "F", ExactSpelling = true)]
     private static extern partial void F(); // Compliant
-}
-
-// https://github.com/SonarSource/sonar-dotnet/issues/7629
-public class Repro_7629
-{
-    interface Interface_7629
-    {
-        void MyMethod();
-        static abstract void MyStaticAbstractMethod();
-    }
-
-    class MyClass_7629 : Interface_7629
-    {
-        public void MyMethod() { } // Noncompliant FP
-        public static void MyStaticAbstractMethod() { } // Noncompliant FP
-    }
 }
